Adding Eslint for project

+++ src/Applications/use_case/DeleteCommentUseCase.js
@@ -1,4 +1,4 @@
-const DeleteComment = require("../../Domains/comments/entities/DeleteComment");
+const DeleteComment = require('../../Domains/comments/entities/DeleteComment');
 
 class DeleteCommentUseCase {
   constructor({ threadRepository, commentRepository }) {
@@ -8,13 +8,17 @@ class DeleteCommentUseCase {
 
   async execute(useCasePayload) {
     this._validatePayload(useCasePayload);
-    await this._threadRepository.getById(useCasePayload.threadId);
+    await this._threadRepository.verifyThreadExists(useCasePayload.threadId);
     await this._commentRepository.verifyCommentExists(useCasePayload.commentId);
-    return await this._commentRepository.deleteComment(useCasePayload);
+    await this._commentRepository.verifyCommentOwner(
+      useCasePayload.commentId,
+      useCasePayload.owner,
+    );
+    return this._commentRepository.deleteComment(useCasePayload);
   }
 
   _validatePayload = (payload) => {
-    new DeleteComment(payload);
+    DeleteComment(payload);
   };
 }
